Added notification for when the backup fails

+++ scripts/backup/backup-forget.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
-restic snapshots -r /home/coffeevector/Backup/resticBackup | head -n -2 | awk '{if(NR!=1&&NR!=2){print $1 " " $2 " " $3}}' > /tmp/restic-snapshots.txt
+output=$(rofi -dmenu -password -lines 0 -p "Password" | restic snapshots -r /home/coffeevector/Backup/resticBackup | head -n -2 | awk '{if(NR!=1&&NR!=2){print $0}}' | tee  /tmp/restic-snapshots.txt)
+echo $output
+if [ "$output" = "" ]; then
+	notify-send "SNAPSHOT FORGET FAILED"
+	exit 1
+fi
 snapshot=$(cat /tmp/restic-snapshots.txt | rofi -dmenu -i -p "Snapshot" | awk '{print $1}')
 rm /tmp/restic-snapshots.txt
-restic forget $snapshot
+if [ "$snapshot" = "" ]; then
+	notify-send "EXITING SNAPSHOT FORGET"
+	exit 1
+fi
+output=$(rofi -dmenu -password -lines 0 -p "Password" | restic forget $snapshot -r /home/coffeevector/Backup/resticBackup)
+if [ "$output" = "" ]; then
+	notify-send "SNAPSHOT FORGET FAILED"
+else
+	notify-send "SNAPSHOT FORGET COMPLETE"
+fi
